Move visualizations-related components to own folder
