+++ client/app/components/dashboards/ExpandedWidgetDialog.jsx
@@ -3,8 +3,8 @@ import PropTypes from "prop-types";
 import Button from "antd/lib/button";
 import Modal from "antd/lib/modal";
 import { wrap as wrapDialog, DialogPropType } from "@/components/DialogWrapper";
-import VisualizationRenderer from "@/visualizations/VisualizationRenderer";
-import VisualizationName from "@/visualizations/VisualizationName";
+import VisualizationRenderer from "@/visualizations/components/VisualizationRenderer";
+import VisualizationName from "@/visualizations/components/VisualizationName";
 
 function ExpandedWidgetDialog({ dialog, widget }) {
   return (
