+++ client/app/visualizations/components/VisualizationName.less
